+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Setup Reliably workflow for CI/CD platform; [44](https://github.com/reliablyhq/cli/issues/44)
 - Show suggestions level (info, warning, error) on discovery result; [52](https://github.com/reliablyhq/cli/issues/52)
 - Discover `--level` flag allows to display only suggestions at specified level or higher
+- Discover `--live` flag looks for weaknesses in a live Kubernetes cluster
 
 ### Changed
 - `discover` command now requires to be authenticated to Reliably.